pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "heatmatch"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "matplotlib",
    "Pillow",
]